+++ test/e2e/app-dir/app/app/hooks/use-pathname/server/page.js
@@ -1,8 +1,8 @@
-import { usePathname } from 'next/dist/client/components/hooks-client'
+// import { usePathname } from 'next/dist/client/components/hooks-client'
 
 export default function Page() {
   // This should throw an error.
-  usePathname()
+  // usePathname()
 
   return null
 }
