+++ test/e2e/app-dir/app/app/hooks/use-router/server/page.js
@@ -1,8 +1,8 @@
-import { useRouter } from 'next/dist/client/components/hooks-client'
+// import { useRouter } from 'next/dist/client/components/hooks-client'
 
 export default function Page() {
   // This should throw an error.
-  useRouter()
+  // useRouter()
 
   return null
 }
