+++ test/e2e/app-dir/app/app/hooks/use-search-params/server/page.js
@@ -1,8 +1,8 @@
-import { useSearchParams } from 'next/dist/client/components/hooks-client'
+// import { useSearchParams } from 'next/dist/client/components/hooks-client'
 
 export default function Page() {
   // This should throw an error.
-  useSearchParams()
+  // useSearchParams()
 
   return null
 }
